Move bulk_statistics test to separate file
